+++ database/migrations/2022_09_22_090854_create_hotel_packages_table.php
@@ -8,6 +8,10 @@ class CreateHotelPackagesTable extends Migration
 {
     public function up()
     {
+        if (Schema::hasTable('hotel_packages')) {
+            return;
+        }
+
         Schema::create('hotel_packages', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->foreignId("hotels_id")->constrained("hotels")->onDelete("cascade");
