+++ database/migrations/2022_12_23_104043_create_transport_prices_table.php
@@ -8,6 +8,10 @@ class CreateTransportPricesTable extends Migration
 {
     public function up()
     {
+        if (Schema::hasTable('transport_prices')) {
+            return;
+        }
+
         Schema::create('transport_prices', function (Blueprint $table) {
             $table->id();
             $table->foreignId("transports_id")->constrained("transports")->onDelete("cascade");
