+++ database/migrations/2025_06_02_163805_create_villa_prices_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('villa_prices')) {
+            return;
+        }
+
         Schema::create('villa_prices', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->foreignId('villa_id')->constrained()->onDelete('cascade');
