+++ database/migrations/2025_09_26_145538_create_carts_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('carts')) {
+            return;
+        }
+
         Schema::create('carts', function (Blueprint $table) {
             $table->id();
             $table->foreignId('users_id')->constrained()->onDelete('cascade'); // user pemilik cart
